Added tests for the GLB Asset property.
Also finished the standard textures slots and materials properties.
I still haven't touched the extensions yet, but I figure I might do
this after dealing with all the other standard properties.

Added tests to check for the "asset" property and to make sure that
the information stored there are available through the loader.

Signed-off-by: Voyage <voyage@miouyouyou.fr>