Added tests for "textures"
Since the GLB Loader was using a stupid shortcut where
"images" were parsed and considered "textures", which is
absolutely not how it should work, but was still working
nicely in "some" cases.

The test allowed me to fix the GLB Loader so that it parses
"samplers" and "textures" informations correctly, loading
the appropriate images into the texture, and using the right
sampling properties.

Signed-off-by: Voyage <voyage@miouyouyou.fr>